Add S6957 (`eslint-plugin-react/no-deprecated`) extracted from S1874 (`deprecation`)
