+++ its/ruling/src/test/expected/jsts/reddit-mobile/javascript-S6957.json
