+++ its/ruling/src/test/expected/jsts/redux/javascript-S6957.json
