+++ packages/jsts/src/rules/S6957/cb.options.json
