+++ packages/jsts/src/rules/S6957/cb.test.ts
@@ -20,9 +20,16 @@
 import { check } from '../tools';
 import { rule } from './';
 import path from 'path';
+import { clearPackageJsons, loadPackageJsons } from '@sonar/jsts';
 
 const sonarId = path.basename(__dirname);
 
-describe('Rule S1874', () => {
+describe('Rule S6957', () => {
+  beforeEach(() => {
+    loadPackageJsons(__dirname, []);
+  });
+  afterAll(() => {
+    clearPackageJsons();
+  });
   check(sonarId, rule, __dirname);
 });
