+++ packages/jsts/src/rules/S6957/fixtures/noreact1/package.json
