+++ packages/jsts/src/rules/S6957/fixtures/noreact2/package.json
