+++ packages/jsts/src/rules/S6957/fixtures/react15/package.json
