+++ packages/jsts/src/rules/S6957/fixtures/react19/package.json
