+++ packages/jsts/src/rules/S6957/index.ts
@@ -17,12 +17,4 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-import { check } from '../tools';
-import { rule } from './';
-import path from 'path';
-
-const sonarId = path.basename(__dirname);
-
-describe('Rule S1874', () => {
-  check(sonarId, rule, __dirname);
-});
+export { rule } from './rule';
