+++ packages/jsts/src/rules/S6957/package.json
