+++ sonar-plugin/javascript-checks/src/main/java/org/sonar/javascript/checks/CheckList.java
@@ -257,6 +257,7 @@ public final class CheckList {
       NoConstructorReturnCheck.class,
       NoControlRegexCheck.class,
       NoDangerWithChildrenCheck.class,
+      NoDeprecatedReactCheck.class,
       NoDirectMutationStateCheck.class,
       NoDuplicateEnumValuesCheck.class,
       NoDuplicateImportsCheck.class,
