+++ sonar-plugin/javascript-checks/src/main/resources/org/sonar/l10n/javascript/rules/javascript/Sonar_way_profile.json
@@ -327,6 +327,7 @@
     "S6854",
     "S6855",
     "S6859",
-    "S6861"
+    "S6861",
+    "S6957"
   ]
 }
